\name{imgs}
\docType{data}
\alias{imgs}
\title{Video Image Frame list as retrieved by ExtractionImgPos}
\description{
Example of vector of image frames available
}
\format{A vector with 30 entries, each a string with the postion of the available frame}
\source{TAMIDA 2013}
\references{
TAMIDA 2013
}
\keyword{datasets}
