\name{VideoComparison-package}
\alias{VideoComparison-package}
\alias{VideoCompare-package}
\docType{package}
\title{
Comparison between videos.
}
\description{
Comparison between videos by their descritors 
(motion curves and image hashes for transitions)
}
\details{
\tabular{ll}{
Package: \tab VideoCompare\cr
Type: \tab Package\cr
Version: \tab 0.13\cr
Date: \tab 2015-05-22\cr
License: \tab Under GPL 2.0+ \cr
Description: Functions to compare video motion vector and image hashes
}
}
\author{
Silvia Espinosa, <silviaespinosag@gmail.com>,
Jose Maria Perez, <josem.perez.ramos@gmail.com>,
Joaquin Ordieres, <j.ordieres@upm.es>,
Antonio Bello, <abello@uniovi.es>.

Maintainer: Joaquin Ordieres <j.ordieres@upm.es>
}
\references{
Espinosa-Gutiez, S., Ordieres-Mere, J., Bello-Garcia, A.: Large scale part-to-part video matching by a 
likelihood function using featured based video representation. TAMIDA 2013 - Taller de Mineria de Datos dentro del IV 
congreso Espannol de Informatica. 254-257 (2013). \url{http://bioinspired.dacya.ucm.es/maeb2013/images/ActasCAEPIA_final.pdf}
\cr
Cui, M., Femiani, J., Hu, J., Wonka, P., Razdan, A.: Curve matching for 
open 2D curves. Pattern Recogn. Lett. 30, 1-10 (2009)
}
\keyword{ package }
